Renaming namespace tools to library
